src/handlers/RegisterDnsZone/requirements.txt
aws_lambda_powertools
boto3-stubs[route53,sts]
crhelper